products/requirements.txt
cffi==1.17.0
cryptography==43.0.0
pycparser==2.22
PyJWT==2.9.0
PyYAML==6.0.2
